Windows : Create and install a startup.json file with all the startup parameters
Crosswalk is expected to launch with command line parameters which will
tell the URL/Manifest file to render as well as all the option (like
debug, extensions and so forth). In an environment where Crosswalk is
expected to launch without command line parameters (such as in stores like
Windows Store) app tools will produce a startup.json with all these informations
and install it. Crosswalk will then read it to launch itself.